# docs/api-docs.yaml
openapi: 3.0.0
info:
  title: API Multiservicios
  description: Backend para gestión de inventario, ventas y productos de tienda multiservicios.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Producto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Coca Cola 3L"
        descripcion:
          type: string
          example: "Bebida gaseosa familiar"
        id_rubro:
          type: integer
          example: 1
        precio_venta:
          type: number
          format: float
          example: 12.50
        stock:
          type: integer
          example: 50
        stock_minimo:
          type: integer
          example: 5
        codigo_barra:
          type: string
          example: "770123456789"
        sku:
          type: string
          example: "BEB-001"
        imagen_url:
          type: string
          nullable: true
          example: "https://storage.googleapis.com/emagenes/productos/producto_123.jpg"
        status:
          type: integer
          description: "1: Activo, 0: Inactivo/Eliminado"
          example: 1
    ProductoInput:
      type: object
      required:
        - nombre
        - id_rubro
        - precio_venta
      properties:
        nombre:
          type: string
        id_rubro:
          type: integer
        precio_venta:
          type: number
        stock:
          type: integer
        stock_minimo:
          type: integer
        codigo_barra:
          type: string
        sku:
          type: string
        descripcion:
          type: string
...
security:
  - bearerAuth: []
paths:
  /api/productos:
    get:
      summary: Listar productos
      description: Obtiene una lista paginada de productos con filtros opcionales.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Número de página
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Productos por página
        - in: query
          name: search
          schema:
            type: string
          description: Buscar por nombre, código de barra o SKU
        - in: query
          name: id_rubro
          schema:
            type: integer
          description: Filtrar por ID de rubro
        - in: query
          name: bajo_stock
          schema:
            type: boolean
          description: Si es true, devuelve solo productos con stock bajo
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Producto'
                  pagination:
                    type: object
                    properties:
                      total: 
                        type: integer
                      page:
                        type: integer
                      pages: 
                        type: integer
    post:
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Producto'
...